Feature PS-9868: Add Job logs to job group in cli
