+++ paperspace/cli/cli.py
@@ -7,7 +7,6 @@ from paperspace import constants, client, config
 from paperspace.cli import common
 from paperspace.cli.common import api_key_option, del_if_value_is_none
 from paperspace.cli.jobs import jobs_group
-from paperspace.cli.logs import logs_group
 from paperspace.cli.models import models_group
 from paperspace.cli.projects import projects_group
 from paperspace.cli.cli_types import ChoiceType, json_string
@@ -1068,7 +1067,6 @@ def version():
 cli.add_command(jobs_group)
 cli.add_command(projects_group)
 cli.add_command(models_group)
-cli.add_command(logs_group)
 
 
 if __name__ == '__main__':
