--- tests/functional/test_logs.py
@@ -1,96 +1,0 @@
-import mock
-from click.testing import CliRunner
-
-from paperspace.cli import cli
-from paperspace.client import default_headers
-from tests import MockResponse, example_responses
-
-
-class TestListLogs(object):
-    URL = "https://logs.paperspace.io/jobs/logs?jobId=some_job_id&line=0"
-    EXPECTED_HEADERS = default_headers.copy()
-    EXPECTED_HEADERS_WITH_CHANGED_API_KEY = default_headers.copy()
-    EXPECTED_HEADERS_WITH_CHANGED_API_KEY["X-API-Key"] = "some_key"
-
-    RESPONSE_JSON_WITH_WRONG_API_TOKEN = {"status": 400, "message": "Invalid API token"}
-    EXPECTED_RESPONSE_JSON = example_responses.LIST_OF_LOGS_FOR_JOB
-    BASIC_COMMAND_WITHOUT_PARAMETERS = ["logs", "list"]
-    BASIC_COMMAND = ["logs", "list", "--jobId", "some_job_id", "--apiKey", "some_key"]
-
-    EXPECTED_STDOUT_WITHOUT_PARAMETERS = """Usage: cli logs list [OPTIONS]
-Try "cli logs list --help" for help.
-
-Error: Missing option "--jobId".
-"""
-
-    EXPECTED_STDOUT = """+Job some_job_id logs--------------------------------------------------------------------+
-| LINE | MESSAGE                                                                         |
-+------+---------------------------------------------------------------------------------+
-| 1    | Traceback (most recent call last):                                              |
-| 2    |   File "generate_figures.py", line 15, in <module>                              |
-| 3    |     import dnnlib.tflib as tflib                                                |
-| 4    |   File "/paperspace/dnnlib/tflib/__init__.py", line 8, in <module>              |
-| 5    |     from . import autosummary                                                   |
-| 6    |   File "/paperspace/dnnlib/tflib/autosummary.py", line 31, in <module>          |
-| 7    |     from . import tfutil                                                        |
-| 8    |   File "/paperspace/dnnlib/tflib/tfutil.py", line 34, in <module>               |
-| 9    |     def shape_to_list(shape: Iterable[tf.Dimension]) -> List[Union[int, None]]: |
-| 10   | AttributeError: module 'tensorflow' has no attribute 'Dimension'                |
-| 11   | PSEOF                                                                           |
-+------+---------------------------------------------------------------------------------+
-"""
-
-    EXPECTED_STDOUT_WITH_WRONG_API_TOKEN = "Invalid API token\n"
-
-    @mock.patch("paperspace.cli.cli.client.requests.get")
-    def test_command_should_not_send_request_without_required_parameters(self, get_patched):
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(cli.cli, self.BASIC_COMMAND_WITHOUT_PARAMETERS)
-        print(result)
-
-        get_patched.assert_not_called()
-        assert result.exit_code == 2
-        assert result.output == self.EXPECTED_STDOUT_WITHOUT_PARAMETERS
-
-    @mock.patch("paperspace.cli.cli.client.requests.get")
-    def test_should_send_valid_get_request_and_print_available_logs(self, get_patched):
-        get_patched.return_value = MockResponse(json_data=self.EXPECTED_RESPONSE_JSON, status_code=200)
-
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(cli.cli, self.BASIC_COMMAND)
-
-        get_patched.assert_called_with(self.URL,
-                                       headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
-                                       json=None,
-                                       params=None)
-
-        assert result.output == self.EXPECTED_STDOUT
-        assert result.exit_code == 0
-
-    @mock.patch("paperspace.cli.cli.client.requests.get")
-    def test_should_send_valid_get_request_when_log_list_was_used_with_wrong_api_key(self, get_patched):
-        get_patched.return_value = MockResponse(json_data=self.RESPONSE_JSON_WITH_WRONG_API_TOKEN, status_code=400)
-
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(cli.cli, self.BASIC_COMMAND)
-
-        get_patched.assert_called_with(self.URL,
-                                       headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
-                                       json=None,
-                                       params=None)
-        assert result.output == self.EXPECTED_STDOUT_WITH_WRONG_API_TOKEN
-        assert result.exit_code == 0
-
-    @mock.patch("paperspace.cli.cli.client.requests.get")
-    def test_should_print_error_message_when_error_status_code_received_but_no_content_was_provided(self, get_patched):
-        get_patched.return_value = MockResponse(status_code=400)
-
-        cli_runner = CliRunner()
-        result = cli_runner.invoke(cli.cli, self.BASIC_COMMAND)
-
-        get_patched.assert_called_with(self.URL,
-                                       headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
-                                       json=None,
-                                       params=None)
-        assert result.output == "Error while parsing response data: No JSON\n"
-        assert result.exit_code == 0
