+++ tox.ini
@@ -26,7 +26,7 @@ deps =
 pip_pre = true
 
 commands =
-    pytest --cov=paperspace --cov-append
+    pytest -vv --cov=paperspace --cov-append
 
 [testenv:check]
 deps =
